requirements.txt
autogen-agentchat
#autogen-core==0.4.0.dev13
#autogen-core[grpc]==0.4.0.dev13
#autogen-ext==0.4.0.dev13
autogen-ext[openai,azure]
openai
azure-identity
PyGithub
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-otlp-proto-http
azure-monitor-opentelemetry
opentelemetry-instrumentation-openai
python-dotenv


